Improve function to batch download videos and audio from YouTube
The function now has --audio and --format options, respectively to download the
audio only from a YouTube video, and to manually set the download format, which
can be any video or audio format found when calling youtube-dl --list-formats.

The audio setting is set to format 250, giving a limited audio quality and reducing bandwidth.